Removed SpathUnescapedKey and added SpathKey that returns SpathEscapedKey object
