+++ src/main/java/com/teragrep/pth10/ast/commands/evalstatement/UDFs/Spath.java
@@ -50,6 +50,7 @@ import com.teragrep.pth10.ast.NullValue;
 import com.teragrep.pth10.ast.TextString;
 import com.teragrep.pth10.ast.UnquotedText;
 import com.teragrep.pth10.steps.spath.SpathEscapedKey;
+import com.teragrep.pth10.steps.spath.SpathKey;
 import org.apache.commons.text.StringEscapeUtils;
 import org.apache.spark.sql.api.java.UDF4;
 import org.slf4j.Logger;
@@ -118,7 +119,7 @@ public class Spath implements UDF4<String, String, String, String, Map<String, S
                 for (Map.Entry<String, JsonElement> sub : jsonElem.getAsJsonObject().entrySet()) {
                     // put key:value to map - unescaping result in case was a nested json string
                     result
-                            .put(new SpathEscapedKey(sub.getKey()).escaped(), new UnquotedText(new TextString(StringEscapeUtils.unescapeJson(sub.getValue().toString()))).read());
+                            .put(new SpathKey(sub.getKey()).escaped().toString(), new UnquotedText(new TextString(StringEscapeUtils.unescapeJson(sub.getValue().toString()))).read());
                 }
             }
             // Manual extraction via spath expression (JSON)
@@ -128,7 +129,7 @@ public class Spath implements UDF4<String, String, String, String, Map<String, S
                 );
                 // put key:value to map - unescaping result in case was a nested json string
                 result
-                        .put(new SpathEscapedKey(spathExpr).escaped(), jsonSubElem != null ? new UnquotedText(new TextString(StringEscapeUtils.unescapeJson(jsonSubElem.toString()))).read() : nullValue.value());
+                        .put(new SpathKey(spathExpr).escaped().toString(), jsonSubElem != null ? new UnquotedText(new TextString(StringEscapeUtils.unescapeJson(jsonSubElem.toString()))).read() : nullValue.value());
             }
             return result;
         }
@@ -163,7 +164,7 @@ public class Spath implements UDF4<String, String, String, String, Map<String, S
                     LOGGER.debug("spath->xpath conversion: <[{}]>", spathAsXpath);
 
                     String rv = (String) xPath.compile(spathAsXpath).evaluate(doc, XPathConstants.STRING);
-                    result.put(new SpathEscapedKey(spathExpr).escaped(), rv.trim());
+                    result.put(new SpathKey(spathExpr).escaped().toString(), rv.trim());
                 }
                 return result;
             }
@@ -171,11 +172,11 @@ public class Spath implements UDF4<String, String, String, String, Map<String, S
                 LOGGER.warn("spath: The content couldn't be parsed as JSON or XML. Details: <{}>", e.getMessage());
                 // return pre-existing content if output is the same as input
                 if (nameOfInputCol.equals(nameOfOutputCol)) {
-                    result.put(new SpathEscapedKey(spathExpr).escaped(), input);
+                    result.put(new SpathKey(spathExpr).escaped().toString(), input);
                 }
                 // otherwise output will be empty on error
                 else {
-                    result.put(new SpathEscapedKey(spathExpr).escaped(), nullValue.value());
+                    result.put(new SpathKey(spathExpr).escaped().toString(), nullValue.value());
                 }
                 return result;
             }
@@ -278,13 +279,13 @@ public class Spath implements UDF4<String, String, String, String, Map<String, S
             }
 
             // if there are multiple columns of the same name, add value to existing column
-            final SpathEscapedKey key = new SpathEscapedKey(colName);
-            if (map.containsKey(key.escaped())) {
-                String existingValue = map.get(key.escaped());
-                map.put(key.escaped(), existingValue.concat("\n").concat(rootNode.getTextContent()));
+            final SpathKey key = new SpathKey(colName);
+            if (map.containsKey(key.escaped().toString())) {
+                String existingValue = map.get(key.escaped().toString());
+                map.put(key.escaped().toString(), existingValue.concat("\n").concat(rootNode.getTextContent()));
             }
             else {
-                map.put(key.escaped(), rootNode.getTextContent());
+                map.put(key.escaped().toString(), rootNode.getTextContent());
             }
         }
